cast qps to float32

+++ cluster-autoscaler/cloudprovider/clusterapi/clusterapi_provider.go
@@ -158,7 +158,7 @@ func BuildClusterAPI(opts config.AutoscalingOptions, do cloudprovider.NodeGroupD
 	if err != nil {
 		klog.Fatalf("cannot build management cluster config: %v", err)
 	}
-	managementConfig.QPS = opts.KubeClientQPS
+	managementConfig.QPS = float32(opts.KubeClientQPS)
 	managementConfig.Burst = opts.KubeClientBurst
 
 	workloadKubeconfig := opts.KubeConfigPath
@@ -167,7 +167,7 @@ func BuildClusterAPI(opts config.AutoscalingOptions, do cloudprovider.NodeGroupD
 	if err != nil {
 		klog.Fatalf("cannot build workload cluster config: %v", err)
 	}
-	workloadConfig.QPS = opts.KubeClientQPS
+	workloadConfig.QPS = float32(opts.KubeClientQPS)
 	workloadConfig.Burst = opts.KubeClientBurst
 
 	// Grab a dynamic interface that we can create informers from
